Extract js types to pink-types
