+++ crates/pink/pink-types/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "pink-types"
+version = "0.1.0"
+edition = "2021"
+description = "Phala's ink! for writing phat contracts"
+license = "Apache-2.0"
+keywords = ["phat-contract", "pink", "ink"]
+
+[dependencies]
+scale = { package = "parity-scale-codec", version = "3.3", default-features = false, features = ["derive"] }
+scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
+
+[features]
+default = ["scale/std"]
+std = ["scale/std", "scale-info/std"]
