+++ crates/pink/pink-types/src/lib.rs
@@ -1,0 +1,6 @@
+
+#![cfg_attr(not(feature = "std"), no_std)]
+
+extern crate alloc;
+
+pub mod js;
